Initial commit: Tree-structure friendly arithmetic coding
This is R&D to provide working primitives for other projects.

See `README.md`.

As with many of my projects the polished initial Git commit happens after it's
been in use for some time.  This one is related to the `dbio` fast storage I/O
project and tree-structured storage, going back to Jan 2008.  It was actually
part of a distributed database, with storage being secondary.